refactor: estilização tela-login
