+++ myFinnances/src/css/login.css
@@ -1,0 +1,74 @@
+*{
+    font-family: Inter, sans-serif;
+}
+
+.logo{
+    width: 80px;
+    align-self: center;
+}
+
+.container {
+    height: 100vh;
+    width: 100%;
+
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+
+}
+
+header{
+    display: flex;
+    flex-direction: column;
+}
+
+header h1{
+    font-size: 36px;
+}
+
+.form{
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+}
+
+input{
+    width: 520px;
+    height: 49px;
+    margin: 12px;
+
+    border-radius: 13px;
+    border: 1px solid #000;
+}
+
+.submit{
+    cursor: pointer;
+    width: 300px;
+    height: 49px;
+
+    align-self: center;
+
+    color: white;
+    background-color: #3bf199;
+    border: none;
+}
+
+label{
+    font-size: 16px;
+
+    padding-left: 12px;
+}
+
+.esqueceu-senha{
+    padding-left: 12px;
+}
+
+.new-account{
+    display: flex;
+    justify-content: center;
+}
+section h2{
+    font-size: 16px;
+    font-weight: normal;
+}
